pyproject.toml
[project]
name = "askforum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=0.8.0",
    "google-search-results>=2.4.2",
    "serpapi>=0.1.5",
]
